requirements.txt
Python==3.5.2
Django==2.0
Djangorestframework==3.8.2